refactor: 검증 로직을 각 메서드가 필요한 파일로 이동
