+++ src/main/kotlin/model/Car.kt
@@ -1,14 +1,12 @@
 package model
 
-import util.Validation
-
 class Car(val name: String) {
 
     var forwardCount = DEFAULT_FORWARD_COUNT
         private set
 
     init {
-        Validation.checkNameLength(name)
+        checkNameLength(name)
     }
 
     fun moveForward(isMove: Boolean) {
@@ -19,5 +17,11 @@ class Car(val name: String) {
 
     companion object {
         private const val DEFAULT_FORWARD_COUNT = 0
+        private const val CAR_NAME_MIN_LENGTH = 1
+        private const val CAR_NAME_MAX_LENGTH = 5
+
+        fun checkNameLength(carName: String) {
+            require(carName.length in CAR_NAME_MIN_LENGTH..CAR_NAME_MAX_LENGTH) { "[ERROR] 자동차 이름은 5자 이하여야 합니다." }
+        }
     }
 }
